requirements.txt
langchain==1.2.0
mcpadapt==0.1.19
mcp==1.24.0
fastapi==0.121.0
langchain-openai==1.1.4
langchain-core==1.2.2
langchain-community==0.4.1
loguru==0.7.3
python-docx==1.1.2
pdfplumber==0.11.6
uvicorn==0.38.0